[Refactor] move Instr tests, start combinding compiled test suites
