+++ tests/test_instr.py
@@ -6,289 +6,11 @@ from mccode_antlr.common.expression import Expr
 from mccode_antlr.instr.orientation import Vector
 
 
-def parse_instr_string(instr_source: str):
-    from mccode_antlr.loader import parse_mcstas_instr
-    return parse_mcstas_instr(instr_source)
-
 
 def make_mcstas_assembler(name: str):
     from mccode_antlr.assembler import Assembler
     from mccode_antlr.reader import MCSTAS_REGISTRY
     return Assembler(name, registries=[MCSTAS_REGISTRY])
-
-
-class TestInstr(TestCase):
-
-    def assertRotationsEqual(self, r1, r2):
-        from mccode_antlr.instr.orientation import Rotation, Matrix
-        self.assertTrue(isinstance(r1, Rotation))
-        self.assertTrue(isinstance(r2, Rotation))
-        d = r1 - r2
-        self.assertTrue(isinstance(d, Matrix))
-        self.assertEqual(round(abs(d), 14), Matrix())
-
-    def _positioning_evaluator(self, instr):
-        """Common positioning checks for `test_read_positioning` and `test_assemble_positioning`"""
-        from mccode_antlr.common import Expr
-        from mccode_antlr.instr.orientation import Vector, Rotation
-        z, o = Expr.float(0), Expr.float(1)
-
-        left = instr.get_component('left')
-        v3 = Vector(z, z, o)
-        self.assertEqual(v3, left.orientation.position())
-        self.assertEqual(Rotation(z, z, -o, z, o, z, o, z, z), left.orientation.rotation())
-        self.assertEqual(left.orientation.rotation(), left.orientation.rotation('coordinates').inverse())
-
-        up = instr.get_component('up')
-        v4 = Vector(o, z, z)
-        self.assertEqual(v4, up.orientation.position() - left.orientation.position())
-        up_orientation = Rotation(z, z, -o, -o, z, z, z, o, z)
-        self.assertRotationsEqual(up_orientation, up.orientation.rotation())
-        self.assertEqual(Vector(o, z, o), up.orientation.position())
-        self.assertRotationsEqual(up_orientation, up.orientation.rotation('coordinates').inverse())
-
-        last = instr.get_component('last')
-        v5 = Vector(z, o, z)
-        self.assertRotationsEqual(up_orientation, last.orientation.rotation())
-        self.assertEqual(v5, last.orientation.position() - up.orientation.position())
-
-    def test_assemble_positioning(self):
-        """Equivalent test to `test_read_positioning` but using an assembled instrument"""
-        assembler = make_mcstas_assembler('orientation_test')
-        origin = assembler.component("origin", "Progress_bar", at=[0, 0, 0])
-        left = assembler.component('left', 'Arm', at=([0, 0, 1], origin), rotate=[0, 90, 0])
-        up = assembler.component("up", 'Arm', at=([0, 0, 1], left), rotate=[-90, 0, 0])
-        assembler.component("last", 'Arm', at=([0, 0, 1], up), rotate=[0, 0, 0])
-        self._positioning_evaluator(assembler.instrument)
-
-    def test_read_positioning(self):
-        """Equivalent test to `test_assemble_positioning` but using a parsed instrument"""
-        instr_source = """DEFINE INSTRUMENT orientation_test() TRACE
-        COMPONENT origin = Arm() AT (0, 0, 0) ABSOLUTE
-        COMPONENT left = Arm() AT (0, 0, 1) RELATIVE PREVIOUS ROTATED (0, 90, 0) RELATIVE origin
-        COMPONENT up = Arm() AT (0, 0, 1) RELATIVE left ROTATED (-90, 0, 0) RELATIVE left
-        COMPONENT last = Arm() AT (0, 0, 1) RELATIVE up
-        END """
-        self._positioning_evaluator(parse_instr_string(instr_source))
-
-    def _simple_position_tests(self, instr, positions: dict):
-        from mccode_antlr.common import Expr
-        from mccode_antlr.instr.orientation import Vector
-        positions = {k: Vector(*[Expr.float(x) for x in v]) for k, v in positions.items()}
-        for instance in instr.components:
-            self.assertEqual(positions[instance.name], instance.orientation.position())
-            self.assertEqual(positions[instance.name], instance.orientation.position('axes'))
-            self.assertEqual(positions[instance.name], instance.orientation.position('coordinates'))
-
-    def test_simple_positioning(self):
-
-        from math import pi, cos, sin
-        instr_source = """DEFINE INSTRUMENT simple_test() TRACE
-        COMPONENT origin = Arm() AT (0, 0, 0) ABSOLUTE
-        COMPONENT slit = Arm() at (0, 0, 10) RELATIVE origin
-        COMPONENT sample = Arm() at (0, 0, 10) RELATIVE slit
-        END"""
-        instr = parse_instr_string(instr_source)
-        positions = {'origin': (0, 0, 0), 'slit': (0, 0, 10), 'sample': (0, 0, 20)}
-        self._simple_position_tests(instr, positions)
-
-        instr_source = """DEFINE INSTRUMENT slightly_more_complex() TRACE
-        COMPONENT origin = Arm() AT (0, 0, 0) ABSOLUTE
-        COMPONENT guide_start = Arm() AT (0.01277, 0, 1.930338) RELATIVE origin ROTATED (0, -0.56, 0) RELATIVE origin
-        COMPONENT guide = Arm() AT (0, 0, 0) RELATIVE guide_start
-        COMPONENT guide_end = Arm() AT (0, 0, 4.33) RELATIVE guide
-        END
-        """
-        instr = parse_instr_string(instr_source)
-        positions = {'origin': (0, 0, 0), 'guide_start': (0.01277, 0, 1.930338), 'guide': (0.01277, 0, 1.930338),
-                     'guide_end': (0.01277 - 4.33*sin(pi/180*0.56), 0, 1.930338 + 4.33*cos(pi/180*0.56))}
-        self._simple_position_tests(instr, positions)
-        pos_hat = (0.006615, 0, 0.999978)
-        pos = instr.components[2].orientation.position()
-        distance = pos.length()
-        vector = pos / distance
-        self.assertAlmostEqual(vector.x, pos_hat[0], 6)
-        self.assertAlmostEqual(vector.y, pos_hat[1], 6)
-        self.assertAlmostEqual(vector.z, pos_hat[2], 6)
-        self.assertAlmostEqual(distance, 1.930380239005777, 6)
-
-        # Combining the position and rotation (reduced) operations should yield the same position
-        comb = instr.components[2].orientation.combine().reduce()
-        self.assertEqual(pos, comb.position())
-
-    def test_copy(self):
-        from mccode_antlr.instr import Instr
-        from mccode_antlr.common import InstrumentParameter, Expr, Value, DataType
-        instr_source = """
-        DEFINE INSTRUMENT test_copy(par0=3.14159, double par1 = 49, int par2 =     1010110
-            , string par3="this is a long string with spaces", 
-            
-            string par4, int par5, double par6, par7)
-        TRACE
-        COMPONENT first = Arm() AT (0, 0, 0) ABSOLUTE
-        COMPONENT second = Arm() AT (0, 0, 1) RELATIVE first ROTATED (0, 90, 0) RELATIVE first
-        END
-        """
-        instr = parse_instr_string(instr_source)
-        instr_copy = instr.copy()
-        self.assertTrue(isinstance(instr_copy, Instr))
-        self.assertEqual(instr.name, instr_copy.name)
-        self.assertEqual(len(instr.parameters), len(instr_copy.parameters))
-        self.assertEqual(len(instr.components), len(instr_copy.components))
-        for p, p_copy in zip(instr.parameters, instr_copy.parameters):
-            self.assertTrue(isinstance(p_copy, InstrumentParameter))
-            self.assertEqual(p.name, p_copy.name)
-            self.assertEqual(p.value, p_copy.value)
-
-        instr_copy.add_parameter(InstrumentParameter('par8', '', Expr.float(1.6189)))
-        self.assertEqual(len(instr.parameters), len(instr_copy.parameters) - 1)
-
-    def test_mcpl_split(self):
-        instr_source = """
-        DEFINE INSTRUMENT test_copy()
-        TRACE
-        COMPONENT origin = Arm() AT (0, 0, 0) ABSOLUTE
-        COMPONENT first = Arm() AT (0, 0, 100) RELATIVE origin ROTATED (0, 90, 0) RELATIVE origin
-        COMPONENT second = Arm() AT (0, 0, 1) RELATIVE first ROTATED (-90, 0, 0) RELATIVE first
-        COMPONENT split_here = Arm() AT (0, 0, 0) RELATIVE second
-        COMPONENT third = Arm() AT (0, 0, 10) RELATIVE second
-        END
-        """
-        instr = parse_instr_string(instr_source)
-        before, after = instr.mcpl_split('split_here', filename='test_mcpl_split')
-        self.assertTrue(isinstance(before, Instr))
-        self.assertTrue(isinstance(after, Instr))
-        self.assertEqual(len(before.components), 4)
-        self.assertEqual(len(after.components), 2)
-        for inst, name in zip(before.components, ('origin', 'first', 'second', 'split_here')):
-            self.assertEqual(name, inst.name)
-
-        for inst, name in zip(after.components, ('split_here', 'third')):
-            self.assertEqual(name, inst.name)
-
-        self.assertEqual(before.components[-1].type.name, 'MCPL_output')
-        self.assertEqual(after.components[0].type.name, 'MCPL_input')
-
-        # It is imperative that the split-point did not move:
-        sp = after.components[0]
-        at_rel = sp.at_relative
-        self.assertTrue(at_rel[1] is None)  # the position is absolute
-        pos = Vector(Expr.float(1), Expr.float(0), Expr.float(100))
-        self.assertEqual(pos, at_rel[0])
-
-        # The final component had a relative position which spanned the split point, so it should now be absolute
-        self.assertTrue(after.components[-1].at_relative[1] is None)
-        # rounding makes third.x 1.000...07
-        for a, b in zip(Vector(Expr.float(1), Expr.float(10), Expr.float(100)), after.components[-1].at_relative[0]):
-            self.assertAlmostEqual(a, b)
-
-    def test_tas1_c1(self):
-        from mccode_antlr.loader.loader import parse_mccode_instr_parameters
-        contents = """ DEFINE INSTRUMENT tas(PHM=-37.077,TTM=-74,C1=30) TRACE END"""
-        instr_parameters = parse_mccode_instr_parameters(contents)
-        self.assertEqual(len(instr_parameters), 3)
-        self.assertEqual(instr_parameters[0].name, 'PHM')
-        self.assertEqual(instr_parameters[1].name, 'TTM')
-        self.assertEqual(instr_parameters[2].name, 'C1')
-        self.assertEqual(instr_parameters[0].value, -37.077)
-        self.assertEqual(instr_parameters[1].value, -74)
-        self.assertEqual(instr_parameters[2].value, 30)
-
-    def test_used_parameter_check(self):
-        from mccode_antlr.assembler import Assembler
-        from mccode_antlr.reader import MCSTAS_REGISTRY
-        assembler = Assembler('test_used_parameter_check', registries=[MCSTAS_REGISTRY])
-        assembler.parameter('double par0/"pi" = 3.14159')
-        assembler.parameter('int par1 = 49')
-        assembler.parameter('int par2 = 1010110')
-        assembler.parameter('string par3 = "this is a long string with spaces"')
-        assembler.parameter('string par4 = "the-fourth_parameter"')
-        assembler.parameter('par5 = -9.11')
-        assembler.parameter('a4/"degree" = 90')
-        assembler.component('origin', 'Progress_bar', at=(0, 0, 0))
-        assembler.component('left', 'PSD_monitor', at=(0, 0, 1), parameters=dict(
-            xwidth='par0', yheight='2*fmod(par5, 0.1)', nx='par1', ny='par2 >> 4', restore_neutron=1))
-        # TODO (maybe) the right shift operator is not implemented, instead the expression is parsed as
-        #      BinOp('>', [BinOp('>', ['par2', '']), '4']) {or similar} which passes this test but may not
-        #      be sufficient for all cases
-        assembler.component('up', 'PSD_monitor', at=(0, 0, 2), parameters=dict(
-            xwidth='cos(DEG2RAD * a4)', yheight='sin(DEG2RAD * a4)', nx='-a4 * DEG2RAD', restore_neutron=1))
-
-        inst = assembler.instrument
-        self.assertEqual(7, len(inst.parameters))
-        self.assertEqual(3, len(inst.components))
-        self.assertTrue(inst.parameter_used('par0'))
-        self.assertTrue(inst.parameter_used('par1'))
-        self.assertTrue(inst.parameter_used('par2'))
-        self.assertFalse(inst.parameter_used('par3'))
-        self.assertFalse(inst.parameter_used('par4'))
-        self.assertTrue(inst.parameter_used('par5'))
-        self.assertTrue(inst.parameter_used('a4'))
-        self.assertEqual(2, inst.check_instrument_parameters(), )
-
-    def test_split_broken_reference(self):
-        from textwrap import dedent
-        instr = dedent("""\
-        DEFINE INSTRUMENT test_tof(phase/"degree"=0)
-        TRACE
-        COMPONENT Origin = Progress_bar() AT (0, 0, 0) ABSOLUTE 
-        COMPONENT ESS_source = ESS_butterfly(
-            sector="W", beamline=4, yheight=0.03, cold_frac=0.5, dist=1, focus_xw=0.01, focus_yh=0.01, 
-            c_performance=1.0, t_performance=1.0, Lmin=1, Lmax=2, tmax_multiplier=1.5, n_pulses=1, acc_power=2.0
-        ) AT (0, 0, 0) ABSOLUTE 
-        COMPONENT guide = Guide_gravity(w1=0.01, w2=0.05, h1=0.01, h2=0.05, l=8.0, m=3.5, G=-9.82) AT (0, 0, 1) ABSOLUTE 
-        COMPONENT guide_end = Arm() AT (0, 0, 8.0) RELATIVE guide
-        COMPONENT chopper = DiskChopper(radius=0.35, nu=14, phase=phase, theta_0=115) AT (0, 0, 0.01) RELATIVE guide_end
-        COMPONENT split_before = Arm() AT (0, 0, 1e-08) RELATIVE chopper
-        COMPONENT split_at = Arm() AT (0, 0, 0) RELATIVE split_before
-        COMPONENT split_after = Arm() AT (0, 0, 0) RELATIVE split_before
-        COMPONENT sample = Incoherent(
-            radius=0.005, yheight=0.02, thickness=0.001, focus_ah=2.0, focus_aw=2.0, target_x=1.0, target_y=0.0, 
-            target_z=0.0, Etrans=0.0, deltaE=0.2
-        ) AT (0, 0, 1) RELATIVE chopper 
-        COMPONENT a4 = Arm() AT (0, 0, 0) RELATIVE sample ROTATED (0, 90, 0) RELATIVE sample
-        COMPONENT analyzer = Arm() AT (0, 0, 1) RELATIVE a4
-        COMPONENT a6 = Arm() AT (0, 0, 0) RELATIVE analyzer ROTATED (0, -90, 0) RELATIVE analyzer
-        COMPONENT detector = Arm() AT (0, 0, 1) RELATIVE a6
-        END""")
-        instr = parse_instr_string(instr)
-        first, second = instr.split('split_at', remove_unused_parameters=True)
-        # The last component in the first part of the instrument _is_ split_at
-        self.assertEqual('split_at', first.components[-1].name)
-        # The position of the split_at component should be absolute and the same as in the main instrument
-        split_at = second.components[0]
-        self.assertEqual(split_at.name, 'split_at')
-        at_ref = split_at.at_relative
-        self.assertTrue(isinstance(at_ref[0], Vector))
-        self.assertTrue(at_ref[1] is None)
-        v = Vector(Expr.float(0), Expr.float(0), Expr.float(1 + 8 + 0.01 + 1E-8))
-        self.assertEqual(v, at_ref[0])
-
-        # The sample _should not_ still depend on the chopper, which is only present in the first instrument!
-        sample = second.components[2]
-        self.assertEqual(sample.name, 'sample')
-        at_ref = sample.at_relative
-        self.assertTrue(isinstance(at_ref[0], Vector))
-        self.assertFalse(isinstance(at_ref[1], Instance))
-        self.assertTrue(at_ref[1] is None)
-        # the position of the sample depends on the source, guide, guide-end, and chopper
-        v = Vector(Expr.float(0), Expr.float(0), Expr.float(1 + 8 + 0.01 + 1))
-        self.assertEqual(v, at_ref[0])
-
-        # But everything after _should_ still depend on relative positions
-        for i, name in enumerate(('a4', 'analyzer', 'a6', 'detector')):
-            # Verify that we positioned the components relative in the initial instrument
-            i_inst = instr.components[len(first.components)+2+i]
-            self.assertEqual(name, i_inst.name)
-            self.assertTrue(i_inst.at_relative[1] is not None)
-            self.assertTrue(i_inst.rotate_relative[1] is not None)
-            self.assertTrue(instr.has_component_named(i_inst.at_relative[1].name))
-            self.assertTrue(second.has_component_named(i_inst.at_relative[1].name))
-            # And that, since they're relative to second-instrument components, they're still relative
-            s_inst = second.components[3+i]
-            self.assertEqual(name, s_inst.name)
-            self.assertTrue(s_inst.at_relative[1] is not None)
-            self.assertTrue(s_inst.rotate_relative[1] is not None)
 
 
 class CompiledTest(TestCase):
